chore(deps): Bump dependencies for upcoming 1.20 release
Bump OpenTelemetry Collector core/contrib from v0.158.0 to v0.160.0,
Prometheus from v0.313.2 to v0.314.0, prometheus/common and
prometheus/client_model to their latest patch releases, and Loki from
v3.7.5 to v3.7.7.

Drop the kgeckhart/prometheus-operator fork in favor of upstream
prometheus-operator v0.93.0, which carries the same rulefmt.Parse fix
the fork existed for.

Adapt to upstream's removal of several deprecated, already-no-op Kafka
client config options (resolve_canonical_bootstrap_servers_only,
auth.sasl.version, group_rebalance_strategy singular,
auth.plain_text/auth.tls) and the k8sattributesprocessor
deployment_name_from_replicaset toggle. Fix a pre-existing bug where
otelcol.receiver.kafka/otelcol.exporter.kafka's
authentication { plaintext { } } block silently dropped its
credentials, by translating it to SASL with mechanism PLAIN instead.
Update component docs accordingly.

+++ internal/component/loki/source/podlogs/relabel.go
@@ -27,7 +27,7 @@ func convertRelabelConfig(in []*promv1.RelabelConfig) ([]*relabel.Config, error)
 			outRule.Regex = regex
 		}
 		if inRule.Modulus != 0 {
-			outRule.Modulus = inRule.Modulus
+			outRule.Modulus = uint64(inRule.Modulus)
 		}
 		if inRule.TargetLabel != "" {
 			outRule.TargetLabel = inRule.TargetLabel
